DSPTables: Move interpreter specifics to DSPInterpreter
