+++ Source/Core/Core/DSP/DSPDisassembler.cpp
@@ -3,6 +3,8 @@
 // Licensed under GPLv2+
 // Refer to the license.txt file included.
 
+#include "Core/DSP/DSPDisassembler.h"
+
 #include <algorithm>
 #include <cstdlib>
 #include <fstream>
@@ -12,7 +14,7 @@
 #include "Common/FileUtil.h"
 #include "Common/StringUtil.h"
 
-#include "Core/DSP/DSPDisassembler.h"
+#include "Core/DSP/DSPInterpreter.h"
 #include "Core/DSP/DSPTables.h"
 
 DSPDisassembler::DSPDisassembler(const AssemblerSettings& settings) : settings_(settings)
@@ -192,9 +194,10 @@ bool DSPDisassembler::DisassembleOpcode(const u16* binbuf, int base_addr, int pa
       break;
     }
   }
-  const DSPOPCTemplate fake_op = {
-      "CW",  0x0000, 0x0000, nop,   nullptr, 1, 1, {{P_VAL, 2, 0, 0, 0xffff}},
-      false, false,  false,  false, false};
+  const DSPOPCTemplate fake_op = {"CW",    0x0000, 0x0000, DSPInterpreter::nop,
+                                  nullptr, 1,      1,      {{P_VAL, 2, 0, 0, 0xffff}},
+                                  false,   false,  false,  false,
+                                  false};
   if (!opc)
     opc = &fake_op;
